Casted return to int

+++ Essentials/src/com/earth2me/essentials/craftbukkit/SetExpFix.java
@@ -49,18 +49,19 @@ public class SetExpFix
 		if (level > 30)
 		{
 			//return 62 + (level - 30) * 7;
-			return 4.5*Math.pow(level, 2) - 162.5*level + 2220;
+			return (int) (4.5*Math.pow(level, 2) - 162.5*level + 2220);
 			
 		}
 		if (level > 15)
 		{
 			//return 17 + (level - 15) * 3;
-			return 2.5*Math.pow(level, 2) - 40.5 * level + 360;
+			return (int) (2.5*Math.pow(level, 2) - 40.5 * level + 360);
 		}
 		//return 17;
-		return Math.pow(level, 2) + (6 * level);
+		return (int) (Math.pow(level, 2) + (6 * level));
 	}
-	
+
+    //todo Update this
 	public static int getExpToLevel(final int level)
 	{
 		int currentLevel = 0;
